function/requirements.txt
azure-functions==1.24.0
azure-data-tables
azure-identity